introduce TimeSlotGenerator & update DataHolder.kt
